Make an effort to use a smaller type for memory addresses (including the PC). Log more if PC will be too large.

+++ src/decodeStage.c
@@ -112,7 +112,7 @@ decodeStage(forwardType *fwd)
  */
 void
 updateDregister(stat_t stat, icode_t icode, unsigned int ifun,
-    rregister rA, rregister rB, unsigned int valC, unsigned int valP)
+    rregister rA, rregister rB, unsigned int valC, unsigned short valP)
 {
 	D.stat = stat;
 	D.icode = icode;
